+++ resources/forward_zone.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
-ForwardZone = Struct.new(:name, :forwarders, :forward)
+ForwardZone = Struct.new(:name, :forwarders, :forward, :type)
 
 property :bind_config, String, default: 'default'
 property :forward, String, default: 'only', equal_to: %w(only first)
@@ -10,11 +10,11 @@ action :create do
     find_resource!(:bind_config, new_resource.bind_config)
   end
 
-  bind_config_template = with_run_context :root do
-    find_resource!(:template, bind_config.conf_file)
+  forward_config_template = with_run_context :root do
+    find_resource!(:template, bind_config.forward_zones)
   end
 
-  bind_config_template.variables[:forward_zones] << ForwardZone.new(
-    new_resource.name, new_resource.forwarders, new_resource.forward
+  forward_config_template.variables[:zones] << ForwardZone.new(
+    new_resource.name, new_resource.forwarders, new_resource.forward, 'forward'
   )
 end