+++ resources/primary_zone_template.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 require 'digest'
 
-PrimaryZone = Struct.new(:name, :options)
+PrimaryZone = Struct.new(:name, :options, :type)
 
 property :bind_config, String, default: 'default'
 
@@ -79,14 +79,14 @@ action :create do
     )
     mode 0o440
     action :create
-    notifies :restart, "bind_service[#{bind_service.name}]", :delayed
+    notifies :reload, "bind_service[#{bind_service.name}]", :delayed
   end
 
-  bind_config_template = with_run_context :root do
-    find_resource!(:template, bind_config.conf_file)
+  master_config_template = with_run_context :root do
+    find_resource!(:template, bind_config.primary_zones)
   end
 
-  bind_config_template.variables[:primary_zones] << PrimaryZone.new(
-    new_resource.name, new_resource.options
+  master_config_template.variables[:zones] << PrimaryZone.new(
+    new_resource.name, new_resource.options, 'master'
   )
 end